+++ database/migrations/2016_09_02_000000_create_bvam_categories_table.php
@@ -30,7 +30,7 @@ class CreateBvamCategoriesTable extends Migration
 
             $table->timestamp('first_validated_at')->nullable()->index();
             $table->timestamp('last_validated_at')->nullable();
-            $table->tinyInteger('confirmations')->nullable()->unsigned();
+            $table->integer('confirmations')->nullable()->unsigned();
 
             $table->unique(['hash', 'version']);
         });
